requirements.txt
Django==4.2.7
celery==5.3.4
redis==5.0.1
paramiko==3.3.1
python-decouple==3.8
Pillow
django-crispy-forms==2.1
crispy-bootstrap5==0.7
gunicorn==21.2.0
channels==4.0.0
channels-redis==4.1.0
asgiref==3.7.2
daphne==4.0.0
django-cors-headers
djangorestframework==3.14.0
drf-spectacular==0.26.5